# api.yaml
{
    "openapi": "3.0.0",
    "info": {
        "title": "Canvas API",
        "version": "1.0",
        "contact": {
            "name": "Xavier Basty-Kjellberg",
            "email": "xavier@hexbee.net"
        },
        "description": "API of the Canvas application"
    },
    "servers": [
        {
            "url": "http://localhost:8800",
            "description": "development"
        }
    ],
    "paths": {
...
        "/v1/{id}/rect": {
            "parameters": [
                {
                    "schema": {
                        "type": "string"
                    },
                    "name": "id",
                    "in": "path",
                    "required": true
                }
            ],
            "post": {
                "summary": "Add rectangle to document",
                "operationId": "post-v1-id-rect",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "rect": {
                                        "$ref": "#/components/schemas/Rectangle"
                                    },
                                    "fill": {
                                        "type": "string",
                                        "minLength": 1,
                                        "maxLength": 1
                                    },
                                    "outline": {
                                        "type": "string",
                                        "minLength": 1,
                                        "maxLength": 1
                                    }
                                },
                                "required": [
                                        "rect"
                                ]
                            },
                            "examples": {
                                "example-1": {
                                    "value": {
                                        "rect": {
                                            "origin": {
                                                "x": 5,
                                                "y": 5
                                            },
                                            "width": 10,
                                            "height": 4
                                        },
                                        "fill": "X",
                                        "outline": "@"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                        "operation"
                ],
                "description": "Draw a rectangle in a document."
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Point": {
                "title": "Point",
                "type": "object",
                "properties": {
                    "x": {
                        "type": "integer"
                    },
                    "y": {
                        "type": "integer"
                    }
                },
                "required": [
                        "x",
                        "y"
                ]
            },
            "Rectangle": {
                "title": "Rectangle",
                "type": "object",
                "properties": {
                    "origin": {
                        "$ref": "#/components/schemas/Point"
                    },
                    "width": {
                        "type": "integer"
                    },
                    "height": {
                        "type": "integer"
                    }
                },
                "required": [
                        "origin",
                        "width",
                        "height"
                ]
            }
        }
    },
...
}